:recycle: Refactor ApplicationsStats filenames to more representative ones
