+++ src/modules/BlockchainApplication/components/ApplicationsStats/ApplicationsStat.styles.js
